Abstract LocationDataSource to meet DIP
