+++ app/src/main/java/com/gamla/mymovies/model/LocationDataSource.kt
@@ -1,0 +1,25 @@
+package com.gamla.mymovies.model
+
+import android.annotation.SuppressLint
+import android.app.Activity
+import android.location.Location
+import com.google.android.gms.location.LocationServices
+import kotlinx.coroutines.suspendCancellableCoroutine
+import kotlin.coroutines.resume
+
+interface LocationDataSource {
+    suspend fun findLastLocation(): Location?
+}
+
+class PlayServicesLocationDataSource(activity: Activity) : LocationDataSource {
+    private val fusedLocationClient = LocationServices.getFusedLocationProviderClient(activity)
+
+    @SuppressLint("MissingPermission")
+    override suspend fun findLastLocation(): Location? =
+        suspendCancellableCoroutine { continuation ->
+            fusedLocationClient.lastLocation
+                .addOnCompleteListener {
+                    continuation.resume(it.result)
+                }
+        }
+}
